requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.6.4
distro==1.9.0
Faker==18.3.1
fastapi==0.109.2
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.39.2
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
gradio==4.17.0
gradio_client==0.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.5
langsmith==0.0.87
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.1
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.2
openai==1.12.0
orjson==3.9.13
packaging==23.2
pandas==1.5.3
pickleshare==0.7.5
Pillow==9.4.0
protobuf==4.25.2
py4j==0.10.9.7
pyarrow==15.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.0.9
pyspark==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.8
pytz==2023.2
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
ruff==0.2.1
seaborn==0.12.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
streamlit==1.31.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.27.0.post1
validators==0.22.0
websockets==11.0.3
yarl==1.9.4
zipp==3.17.0
